Add user management (except adding) on Darwin

+++ types/user.sh
@@ -11,17 +11,21 @@ shell=$(arguments get shell $*)
 groups=$(arguments get groups $*)
 preserve_home=$(arguments get preserve-home $*)
 
-target_platform=$(get_baking_platform)
-
 user_get () {
-  row=$(bake cat /etc/passwd | grep -E "^$1:")
-  stat=$?
+  if baking_platform_is "Linux"; then
+    row=$(bake cat /etc/passwd | grep -E "^$1:")
+    stat=$?
+  elif baking_platform_is "Darwin"; then 
+    row=$(bake dscl . -list /Users | grep -E "^$1$")
+    stat=$?
+  fi
   echo $row
   return $stat
 }
 
 user_shell () {
-  current_shell=$(echo "$1" | cut -d: -f 7)
+  baking_platform_is "Linux" && current_shell=$(echo "$1" | cut -d: -f 7)
+  baking_platform_is "Darwin" && current_shell=$(bake dscl . -read "/Users/$1" UserShell | sed -E 's/^UserShell: //g')
   if [ "$current_shell" != $2 ]; then
     echo $current_shell
     return 1
@@ -31,9 +35,7 @@ user_shell () {
 
 user_groups () {
   current_groups=$(bake groups $1)
-  case $target_platform in
-    Linux) current_groups=$(echo "$current_groups" | cut -d: -f 2) ;;
-  esac
+  baking_platform_is "Linux" && current_groups=$(echo "$current_groups" | cut -d: -f 2)
   missing_groups=
   expected_groups=$(IFS=','; echo $2)
 
@@ -57,8 +59,14 @@ case $action in
     echo "--groups=admin,deploy"
     ;;
   status)
-    needs_exec "useradd" || return $STATUS_FAILED_PRECONDITION
-    needs_exec "userdel" || return $STATUS_FAILED_PRECONDITION
+    if baking_platform_is "Linux"; then
+      needs_exec "useradd" || return $STATUS_FAILED_PRECONDITION
+      needs_exec "userdel" || return $STATUS_FAILED_PRECONDITION
+    elif baking_platform_is "Darwin"; then
+      needs_exec "dscl" || return $STATUS_FAILED_PRECONDITION
+    else
+      return $STATUS_UNSUPPORTED_PLATFORM
+    fi
 
     row=$(user_get $handle)
     [ "$?" -gt 0 ] && return $STATUS_MISSING
@@ -83,11 +91,16 @@ case $action in
     return 0 ;;
 
   install)
-    args="-m"
-    [ -n "$shell" ] && args="$args --shell $shell"
-    [ -n "$groups" ] && groups_list=(${groups//,/ }) && args="$args --groups $groups"
-    [[ -n "$groups_list" && "${groups_list[0]}" == "$handle" ]] && args="$args -g $handle"
-    bake useradd $args $handle
+    if baking_platform_is "Linux"; then
+      args="-m"
+      [ -n "$shell" ] && args="$args --shell $shell"
+      [ -n "$groups" ] && groups_list=(${groups//,/ }) && args="$args --groups $groups"
+      [[ -n "$groups_list" && "${groups_list[0]}" == "$handle" ]] && args="$args -g $handle"
+      bake useradd $args $handle
+    elif baking_platform_is "Darwin"; then
+      # TODO add user on Darwin
+      echo "not yet implemented"
+    fi
     ;;
 
   upgrade)
@@ -99,16 +112,30 @@ case $action in
     if [ "$?" -gt 0 ]; then
       groups_to_create=$(IFS=','; echo $missing)
       for group in $groups_to_create; do
-        bake adduser $handle $group
+        if baking_platform_is "Linux"; then
+          bake adduser $handle $group
+        elif baking_platform_is "Darwin"; then
+          bake sudo dseditgroup -o edit -a $handle -t user $group
+        fi
       done
     fi
     ;;
 
   remove)
     if [ -z "$preserve_home" ]; then
-      bake userdel -r $handle
+      if baking_platform_is "Linux"; then
+        bake userdel -r $handle
+      elif baking_platform_is "Darwin"; then
+        home_directory=$(bake dscl . -read "/Users/$handle" NFSHomeDirectory | sed -E 's/^NFSHomeDirectory: //g')
+        bake sudo dscl . -delete "/Users/$handle"
+        bake rm -r "$home_directory"
+      fi
     else
-      bake userdel $handle
+      if baking_platform_is "Linux"; then
+        bake userdel $handle
+      elif baking_platform_is "Darwin"; then
+        bake sudo dscl . -delete "/Users/$handle"
+      fi
     fi
     ;;
